refactor(backend): introduce DTOs for API response formatting
