+++ packages/backend/src/app.js
@@ -1,6 +1,6 @@
 import express from 'express';
 import cors from 'cors';
-import { errorHandler } from './middlewares/errorHandler.js';
+import { errorHandler } from './middleware/errorHandler.js';
 
 const app = express();
